requirements.txt
discord.py
PyNaCl
youtube_dl